rename github folder
